pyproject.toml
[project]
name = "pacperpus"
version = "0.1.0"
description = "Pacmann perpustakaan for learning and exercise"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faker>=30.8.2",
    "fastapi[standard]>=0.115.4",
    "kagglehub>=0.3.4",
    "numpy>=2.1.3",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.36",
]
